Refactor command handling in main.cpp
Replaced the previous command searching method from ifs statements to a dictionary using an unordered_map.
Avoids the program having to go through multiple ifs to just search for the entry using the precommand as the key value, giving it a search time of O(1) instead of a linear search time. This also adds a case if there is no key found; therefore, that command does not exist.

Also improves maintainability and scalability, making it easier to implement future commands.